requirements.txt
Django==4.2.7
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-ckeditor==6.7.0
python-dotenv==1.0.0
django-compressor==4.4
django-cleanup==8.0.0
